+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * New FA icons
 * Fix loading editor styles in Gutenberg iframe in 6.3 and newer
 * Fix no block defaults
+* ACF add loadJson function (load from parent instead of child etc.)
 
 ## 0.5.4 (2023-06-26)